chore: move composer-repo.ts to shop module root and fix formatting
